src/main/java/com/order_manager/main/controller/OrderController.java
package com.order_manager.main.controller;

import java.util.List;

import com.order_manager.main.model.Order;
import com.order_manager.main.repository.OrderRepository;

public class OrderController {

    private final OrderRepository orderRepository;

    public OrderController(OrderRepository orderRepository) {
        this.orderRepository = orderRepository;
    }

    public List<Order> getAllOrders() {
        return orderRepository.findAll();
    }

    public Order getOrderById(String id) {
        return orderRepository.findById(id);
    }

    public void createOrder(Order order) {
        orderRepository.save(order);
    }

    public void updateOrder(Order order) {
        orderRepository.update(order);
    }

    public void deleteOrder(String id) {
        orderRepository.deleteById(id);
    }
}
